requirements.txt
Django>=2.2.7,<2.3.0
djangorestframework>=3.10.3,<3.11.0
psycopg2>=2.8.4,<2.9.0
Pillow>=6.2.1,<6.3.0
flake8>=3.7.9,<3.8.0